+++ CHANGELOG.md
@@ -7,9 +7,19 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [1.0.1] - 2025-05-13
+
+### Fixed
+
+- Fix limits default value in mojos
+- Fix sourceBinding default value in mojos
+- Fix empty label in MarkdownFormat
+- Fix missing BOM module
+
 ## [1.0.0] - 2025-05-12
 
 Initial release.
 
-[Unreleased]: https://github.com/nbbrd/nbbrd-maven-tools/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/nbbrd/nbbrd-maven-tools/compare/v1.0.1...HEAD
+[1.0.1]: https://github.com/nbbrd/nbbrd-maven-tools/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/nbbrd/nbbrd-maven-tools/compare/develop...v1.0.0